+++ py/desitarget/mtl.py
@@ -142,12 +142,12 @@ def make_mtl(targets, obscon, zcat=None, trim=False, scnd=None):
         # If the object is confirmed to be a tracer QSO, then don't request more observations
         if (obsconditions.mask(obscon) & obsconditions.mask("DARK")) != 0:
             if zcat is not None:
-                ii = ztargets['SPECTYPE']=='QSO'
-                ii &= (ztargets['ZWARN'] == 0 ) 
-                ii &= (ztargets['Z']<2.1)
-                ii &= (ztargets['NUMOBS']>0)
+                ii = ztargets['SPECTYPE'] == 'QSO'
+                ii &= (ztargets['ZWARN'] == 0)
+                ii &= (ztargets['Z'] < 2.1)
+                ii &= (ztargets['NUMOBS'] > 0)
                 ztargets['NUMOBS_MORE'][ii] = 0
-            
+
     # ADM assign priorities, note that only things in the zcat can have changed priorities.
     # ADM anything else will be assigned PRIORITY_INIT, below.
     priority = calc_priority(targets_zmatcher, ztargets, obscon)